requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.2
django-filter==23.1
djangorestframework==3.14.0
idna==3.4
joblib==1.2.0
Markdown==3.4.3
numpy==1.24.2
pandas==2.0.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
redis==4.5.4
requests==2.28.2
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
spotipy==2.23.0
sqlparse==0.4.3
threadpoolctl==3.1.0
tzdata==2023.3
urllib3==1.26.15
